IR Receiver failing msg
Added some error messages when IR receiver is not enabled but user tries to access it. The messages are showing up in the terminal window so I'm not entirely happy with that. But it's better than crashing and will help customer support. If only we had a way of sending error messages back to Scratch...